Add IndividualPlace page
This page can be navigated to by clicking on a location in Mis Terras. Instead of making API calls, it relies mainly on data passed from page to page with React-router-dom, specifically using Link and Location. This technique still needs to be implemented to improve the performance of IndividualStory. I should also add an API safety catch in case a user wants to manually enter a url an enter a page that way, which would result in no data being passed to the dom.